+++ src/engine/chipUtils.h
